+++ lib/shrine/plugins/logging.rb
@@ -44,6 +44,10 @@ class Shrine
     # Logging is by default disabled in tests, but you can enable it by setting
     # `Shrine.logger.level = Logger::INFO`.
     module Logging
+      def self.load_dependencies(uploader, *)
+        uploader.plugin :hooks
+      end
+
       def self.configure(uploader, logger: nil, stream: $stdout, format: :human)
         uploader.opts[:logging_logger] = logger
         uploader.opts[:logging_stream] = stream
@@ -78,19 +82,19 @@ class Shrine
       end
 
       module InstanceMethods
-        def store(io, context = {})
+        def around_process(io, context)
+          log("process", io, context) { super }
+        end
+
+        def around_store(io, context)
           log("store", io, context) { super }
         end
 
-        def delete(io, context = {})
+        def around_delete(io, context)
           log("delete", io, context) { super }
         end
 
         private
-
-        def processed(io, context = {})
-          log("process", io, context) { super }
-        end
 
         # Collects the data and sends it for logging.
         def log(action, input, context)